rafactor: rename and move util/dagger
