+++ pkg/cmd/up.go
@@ -14,6 +14,7 @@ import (
 	"github.com/h8r-dev/heighliner/pkg/stack"
 	"github.com/h8r-dev/heighliner/pkg/state"
 	"github.com/h8r-dev/heighliner/pkg/util"
+	"github.com/h8r-dev/heighliner/pkg/util/dagger"
 )
 
 func newUpCmd() *cobra.Command {
@@ -96,7 +97,7 @@ func newUpCmd() *cobra.Command {
 			newArgs = append(newArgs, "--no-cache")
 		}
 		err = util.Exec(
-			util.Dagger,
+			dagger.GetPath(),
 			newArgs...)
 		if err != nil {
 			lg.Fatal().Err(err).Msg("failed to execute stack")
